Add BufferPool implementation to the activator and queue-proxy handler. (#6429)
* Add BufferPool implementation to the activator handler.

This adds pooling of the buffers needed to reverse proxy every request from the activator to reduce the burden on memory allocation and likewise on garbage collection.

Note: This implementation currently does not explicitly clean the slices as they are returned to the pool. The ReverseProxy's implementation of copyBuffer doesn't seem to require a cleaned buffer to be returned.

* Add the pool to the queue-proxy as well.

* Add remark on usage safety.