Add a window total rolling computation. (#6447)
* Add a window total rolling computation.

This is the next step in making autoscaler constant time operation.
Currently to compute the total concurrency over the window, we're executing an O(N) operation in seconds, where
N is window width.
Obviously with larger windows (and we do not restrict them right now) this becomes a noticeable operation (given that it's
executing per revision as well).

So now we're computing the running total and sometimes subtract the holes, should the happen (rare occurrence).
Usage is in the follow up PRs.

* fix comments