Fix the multiscaler test flake (#6486)
The flake was happening because the tick interval was quite short (5ms) and by the time
the test progressed under heavy load (some boskos are worse than others) the timeout trigger also fired.
So since both channels are active, a random one is picked, which is sometimes the wrong one.
Or the other way around due to scheduling the test timed out faster than timer would be scheduled.

Anyway, I replaced the real timer with the mock one and now I can control when we fire
the message.

```
>rt -run=TestMultiScalerScaling -count=20
PASS
ok      knative.dev/serving/pkg/autoscaler      5.070s
```