Add more logging to the destroy pod test
We sometimes see 2, but don't see the pod states that are being destroyed.
Current thinking is that it's a pod that got restarted for some reason (since we do not check the states).
In addition, the log inspection shows that we at no point asked in the autoscaler for more than a single pod
even when for panic mode.

/assign @tcnghia