Update the total computation during window update. (#6521)
The previous change missed the possible edge case that total sum needs to be recomputed
whenever the window changes.
This problem might occur only when window shrinks and therefore average can never to to 0
and scale to 0 will never occur, which is a problem.

Why do we permit online updates to the window width is a different question, which
we might want to revisit in future, but right now it's there...