Create SKS before KPA in autoscaler GlobalResync test.
This test runs a "proper" controller process so we don't control when a reconcilation is kicked off. Creating the KPA first therefore leads to a race where reconcilation might happen before the test creates a dummy SKS with a dummy PrivateServiceName filled in. That will never resolve itself because we don't run an SKS controller here so nothing will fill in the missing PrivateServiceName and no decider will ever be created.

Creating the SKS first should deterministically solve this as the reconciler should now see the SKS we created in any case.