Decouple scale-from-0 decisions from arbitrary floating point precisions.
We recently had a regression that caused revisions to never scale to 0 because of the way we calculate the averages via addition and substraction of floating point numbers. Due to their inaccuracy, we ended up in a situation where a value could be very close to 0, but not actually 0. We fixed this in #6573 by rounding all values to 3 decimals.

That however breaks use-cases where a revision has occasionally very short requests. A single request that takes 100 microseconds for examply would report as a concurrency of 0.0001 from the queue-proxy and thus it would get cut off in average computation and the revision would incorrectly be scaled to 0.

This fixes this by making the code not rely on arbitrary precision decisions. Instead we store the sum of ceiled wholes of all values as well. This value will only drop to 0 (and exactly 0) if all values in the window are 0 (and exactly 0).