Wait for reconcilation events in nscert tests. (#6578)
These tests used to call `Reconcile` through the backdoor by abusing the ability to have the controller object available. While that's usually not a big deal, this can cause races between actual reconcilation and the "manual" reconcilation triggered by the tests themselves. This has become most apparent in the failures of nscert.Failure where the configmap watcher of `TestDomainConfigExplicitDefaultDomain` triggered after that test was actually done.

This removes all manual calls to `Reconcile` and instead makes the whole flow dependent on events coming from "proper" `Reconcile` calls. That should make sure that the state of the test controller is consistent and that all updates are processed in proper order and actually done once the test exits.