Generalize precision func and raise default to 6 digits (#6585)
* Generalize precision func and raise default to 6 digits

Since there may be edgecases when 3 digits is not enough.
e.g. big window (1hr) and 1 request, or there might be many pods and just one request
e.g. after a big drop off in traffic.
Hence raise it to 6 digits, which should be correct for all intents and purposes.

* Fix test name

* Use const