Use static test setup for static config tests in nscert.
These two tests don't need to be running inside an entire controller. The old test setup caused a global resync to run in parallel with a normal reconcilation causing the test to flake.

Using a static setup will not run that global resync at all and it's also not needed for this test. Other tests shouldn't be vulnerable to this as they actually change the config in the test itself and rely on said global resync to trigger reconciles.