Some pre-spring cleaning in the scaler code
Main change is that I removed the return value from the apply scale. The way it was written it gives
the hint that internally the desiredScale might be modified. But it is not. Thus no reason to return
the value.
I've updated some comments to match the current reality, changed wording in some, and formatted others.
Rest are just micro changes.