Remove arbitrary Try timeout in the activator. (#6720)
This might be a relict of earlier times, when we had a lot more going on on the critical path in the activator. Now this timeout should be redundant and the request's timeout should actually decide whether or not we want to stop trying to get to the application itself.