the revision controller's child informers now ignore resource version (#6825)
The revision controller will reconcile revisions when child resources
change. This is accomplished by having informers for the child resource
types and filtering them based on the controlling owner.

The filtering of the owner would take into account it's version
(which is included in the apiVersion). Now that we support different
versions of our resources this PR relaxes the version constraint by
only comparing Group & Kind.