Return 504 on error and other sweeps
There exists a more proper response code than 503 to report a timeout
that is 504. QP is a proxy and hence this code is more than suiting for this use case.
We do not guarantee which reponse code is going to be returned in this case (or I could not find).

Other minor nits and bits