Fix timing issues in pool unit tests. (#6883)
These tests have been prone to timing issues. This should solve this by concurrently scheduling all requests, including the one we expect to finish first. All requests will be held by the barrier until the one we expect has finished, as indicated by the context being closed. We can then safely allow all other requests to finish and verify that the error indeed was the one that was returned by the request we expected to finish first, even if it wasn't scheduled first.