Delete the Serving stats reporter. (#6939)
The StatsReporter we have today is used in exactly one place, and the metric it reports doesn't actually make any sense.  If I were to deploy a Service and 3 years later update it with a bad change, then quickly fix it, this metric would actually report that it took the Service *3 years* to become ready.  To make matters worse, somehow the test code for this has crept its way into knative.dev/pkg, so another PR will attempt to clean that up.