Fix the problem with double report from Activator. (#6954)
When scaling from 0 (or after inactivity) we might report twice within
the same second. All subsequent requests reports will report the full
concurrency on the activator.
Due to how we average buckets in the autoscaler, existing behavior
will cause doublecounting in most cases and we'll scale to more pods
than required. E.g. with cc=1 and tu=0.7 if they end up in the same
bucket in autoscaler we'll have 3 pods rather than 2.