Clarify the usage of the map vs constant 1 (#6992)
Performed the analysis and documented, why 1 is a valid choice in the reporting,
which is better than previous misleading notion of map[key] value, which indicated
that a different from  value is possible.
As currently code is written it is impossible.
In the process I found another bug and documented it with a unit test that will start
failing when the #6991 is fixed.