Deflake the kpa test by waiting for KPA to propagate.
https://prow.knative.dev/view/gcs/knative-prow/logs/ci-knative-serving-continuous/1233454607657603073

It shows as a flake and it can be seen in the logs
that the reason for that is because of the SKS creation race.
By waiting for SKS to show up in the informer we resolve the flake.
See #7099

```
>rt -count=10 -run=TestControllerSy
PASS
ok      knative.dev/serving/pkg/reconciler/autoscaling/kpa      1.955s
```