Fix race condition in concurrency metrics test.
Once the reporter is kicked through the report-channel there is no guarantee that the metric has actually been written. Instead, we need to await the stats being written to the stat-channel (which happens after the metrics are written), to guarantee that we'll always see the published metrics. We need to wait for two stats as the first one is eagerly written to make scale-from-0 as quick as possible.