Make sure all probing goroutines are finished on calling Stop(). (#7196)
* Make sure all probing goroutines are finished on calling Stop().

Some of our tests use a managed prober, which essentially polls forever until instructed to stop. We're not really checking that they are actually done after Stop() returns, which causes racy behavior especially around loggers.

This changes that and makes sure that we've received a done signal from all goroutines before we continue.

This also switches our abuse of the spoofing client's `Poll` function to a more explicit usage of an infinite loop to make clear what's going on.

* Prettier version with errgroup.

* Fix pointers.

* Fix feedback.