Stabilize revision global resync test
I have seen a few failures on my machine so far.
One is pretty obvious: we wait 1s for hoooks, while the jitter in the queue for global resync
permits values up to 2s. So cap it at 3 instead.
The other, I saw just once, if when the object to be reconciled did not even
reconcile by the time we started updating CM.