Take capacity straight from the individual podTrackers if available.
If we can do podIP based routing, we know exactly how many requests we sent where and how much concurrency should already be in use on which pod. We also have a smart algorithm that divides pods between activators.

This makes use of that algorithm for the outer concurrency as well to make sure both values are consistent.