Fix race condition in TestActivatorsIndexUpdate. (#7335)
Once the activatorIndex is computed we can not be sure that all the other fields are readily computed yet.

This switches this test to polling the breaker capacity as that will be the last update when recomputing an update from the informers. Once the capacity is where we expect it to be, it's safe to assert all the other fields statically without polling.