Fix connection refused for testvisibility and testvisibilitysplit
When using contour, if there are no publicly facing services (kingress)
connections to port 80 of the public ingress will fail with a connection
refused.

We didn't dig into why contour is doing this, but this change is to
ensure that the public facing service is ready before we test if the
private services are truly private.

Signed-off-by: Shash Reddy <shashwathireddy@gmail.com>