Infinitely wait for a pod to have reserve capacity.
By the moment we reach the `pickPod` function, we already passed the outer semaphore and thus know that capacity is available. Individual requests still race to the respective pods so it's possible to iterate over the entire pod list and not find a pod with capacity even though we passed the outer semaphore.

We can infinitely retry to reconcile that. This also statically proves the encoded comment.