Fix connection refused for testvisibility and testvisibilitysplit (#7343)
* Fix connection refused for testvisibility and testvisibilitysplit

When using contour, if there are no publicly facing services (kingress)
connections to port 80 of the public ingress will fail with a connection
refused.

We didn't dig into why contour is doing this, but this change is to
ensure that the public facing service is ready before we test if the
private services are truly private.

Signed-off-by: Shash Reddy <shashwathireddy@gmail.com>

* Add check for dial errors if we are expecting it to happen

Signed-off-by: Shash Reddy <shashwathireddy@gmail.com>

* Switch to expectation functions

Still need to change the name of function.

Signed-off-by: Shash Reddy <shashwathireddy@gmail.com>

* Another attempt at implementing expecations

- Trying out Andrew's idea for spliiting expectations
to visualize code readability

* Refactor error expectations

- update the code to allow bool parameter instead of passing error
expecations.

Signed-off-by: Andrew Su <asu@pivotal.io>

* Remove aliased import and updated code

Signed-off-by: Shash Reddy <shashwathireddy@gmail.com>

* revert moving of check for private route in test

Co-authored-by: Shash Reddy <shashwathireddy@gmail.com>