Harden the SKS Global resync test even more. (#7447)
Before we were using hooks to verify the object creation, but this was not enough,
since listers would not propagate in time and we'd not get the proper sequence of events.
This way I am waiting for the listers to actually receive all 4 endpoints before we continue on
updating the activator endpoints.

Fixes #7263