Implement subsetting function in SKS (#7531)
* Implement subsetting function in SKS

This has some non trivial bits to implement linear filtering, vs quadric,
so I've pulled that into a separate PR.
The current implementation preserves the order of IP addresses and subsets
in the endpoints object, but with slightly more complexity we can implement
slightly more effective solution. I'll leave it as an excercise for future,
since it is not going to give us much benefit, because the objects are not
pointers, but rather values, which means swapping from tail would incur a copy.

* review