CleanupOnInterrupt() call multiple cleanup functions only once (#7417)
* CleanupOnInterrupt() call multiple cleanup functions only once

When CleanupOnInterrupt() is called multiple times in one test, the
clean up function is called only once.
(You can quickly confirm that CleanupOnInterrupt calls os.Exit(1) here)

Due to this, the TestIstioProbing does not clean up resources or restore Gateway when
it is interrupted.

To fix it, this PR changes CleanupOnInterrupt() to call multiple
cleanup functions only once

* Run clean up function in reverse order

* Use array of func