Quadruple number of samples during traffic split. (#7564)
This takes us from a 1:7000 change in tests to one in which test
flakiness is incredibly unlikely. We can use the binomial distribution
CDF to increase the fraction required if we want higher confidence
that the effective traffic split is likely being applied.

(Thanks to @tcnghia for helping me with the math here :)