Wait for the route propagation to kill the flakes. (#7610)
Same as in other global resync tests, we are not waiting the for the object we want to operate on
to be propagated throughout, which leads to the CM update not operate on anything and hooks time out.

E.g. https://prow.knative.dev/view/gcs/knative-prow/pr-logs/pull/knative_serving/7607/pull-knative-serving-unit-tests/1250448708672163840
Also happens on CI.