Wait for status update in global resync test.
Waiting for the route to be queriable is too weak of a signal here. It only makes sure that the route we manually created is successfully seen by the informer.

Instead we make sure we can find the route AND wait for the routes status to not be empty. That makes sure that the route got reconciled at least once which is the signal we want for the sake of this test. In essence, that makes sure that the Hook is really triggered by the global resync and not by a racy reconcilation, that raced the global resync.