Add #activator dimension to the load test.
It's interesting to see here and to correlate if different runs that got different pod suggestions
and as such different number of activators (I've seen 3-6 range) depending on what autoscaler scraped
(mostly consistent with always case, though).
The test runs are in progress. I'll update prod config when this is lgtm