Fix an issue with pods becoming non-ready. (#7638)
* Fix an issue with pods becoming non-ready.

The user-container might be come non-ready for multitude of reasons, also
QP under load sometimes fail probes and are moved into the  non-ready list.
If we don't reprobe those pods we'll presume they are ready and continue sending
traffic from activator, which is not correct (especially if user-container stopped being ready,
e.g. it lost DB connection).

This PR fixes the situation by updating the healthy set, removing the pods that moved
from ready to the non ready set, which would ensure they are probed.

* fix the suggestion