Add enable-certificate-owner-ref option to cert-manager's controller (#7366)
When running autoTLS and cert-manager, the secrets created by
cert-manager is not cleaned up. Due to this, switching issuer during
e2e such as from selfsigned to http01 could mess up the certficate
especially ns-cert.

By using this option, the secret has ownerref to certificate so it
will be cleaned up when deleting kcert.