Stabilize Activator HA tests (#7580)
* Run at least 100 requests in activator HA test

* this will tolerate 1 failure in 100 requests, without doing this the
overall number of requests can be lower and 1 failure will mean less
than 0.99 successful responses and the test will fail

* Adjust HA test suite timeout to 15 minutes and -failfast

* make sure we fail immediately after first failure and make a dump of
the services; we don't want to run more tests as they manipulate
controller/autoscaler pods and this can hide the actual status of
services at the time of the first failure

* ActivatorHA tests - wait for Controller to elect a new leader

* We enable leader election and scale up HA-ed components right before
we start Go tests. It might happen that the new leader is not yet
elected when we start creating first knative services and the first
knative service is never reconciled correctly. The activator HA test is
usually started first so it is affected most.

* Move waiting for leader Controller to e2e-tests.sh

* Assert service eventually works in ActivatorHA test

* it is not guaranteed that after changing endpoints the ingress will
know about them and the application will be immediately available, we
need to wait until the app is available and serving the corect response

* fixes https://github.com/knative/serving/issues/7566 - the failure from
April 20

* Use kubectl scale instead of patch for HA tests

* the flake from
https://prow.knative.dev/view/gcs/knative-prow/logs/ci-knative-serving-continuous/1251815466259714049
showed that even though the patch to increase the number of replicas to
2 was applied, the spec.replicas of the deployment stayed different;
let's use a more standard command for scaling the deployment

* Remove unused assertServiceWorksNow