Harden parallel probing test.
This test is prone to timing issues as it relies on timing to be able to count collapsed probes into just one probe.

This changes that and gives the test more breathing room. It only asserts that the actual probe counts we see is lower than the amount of probes we asked for. After all, that's what we want to test and should dilude timing issues enough to make this test harder.