Fix flake in TestDestroyPodWithRequests (#7704)
This patch decrease the sleep time in TestDestroyPodWithRequests.

The test occasionally fails due to the incomplete of last
request. This is because the current sleep time (15s + 6s) is longer
than drainSleepDuration so the request was drained. Please refer to
the log:

https://prow.knative.dev/view/gcs/knative-prow/logs/ci-knative-serving-istio-1.5-no-mesh/1253089832167542785

To fix it, this decrease the sleep time for request from 15s to 12s.

/lint