Improve some debug logging. (#7706)
I've been debugging some autoscaler/SKS failures in the e2e tests
and I noticed we're missing #activatrs in the logs.
In addition the Spew'd argument is evaluated each time we're doing debug log
even if debug is not enabled, so gate that, since it might be pretty expensive
(desugar, otoh, is quite cheap and advertised as such).
Finally log (debug, gated) the chosed subset of the endpoints.