Simplify statistic calculation in the queue-proxy. (#7717)
* Simplify statistic calculation in the queue-proxy.

This is a first step towards sharing this code between the queue-proxy and the activator. If the reporting period is known (which it always is in practice) we can calculate the averages over the window without having to keep a lot of intermittent state in maps and such.

While not crucial for the queue-proxy necessarily, we want to avoid that space complexity when applied to the activator as that will have #revisions times that space complexity.

* Roll back channel changes, make computation more generic.

* Comment nits.

* Some test file nits.