We have to wait for initial reconcile as well in route.
So https://storage.googleapis.com/knative-prow/pr-logs/pull/knative_serving/7752/pull-knative-serving-unit-tests/1255248581837721600/build-log.txt this failure
showed that we updated the CM when the initial reconcile was still underway (with old ctx set to the previous value).
Since the check is for status to change it did change but not to what we want.
So we need to wait in both places: for initial and final reconciliations to finish.