Fix flake in TestPathAndPercentageSplit
This patch relaxes the tolerance for TestPathAndPercentageSplit.

TestPathAndPercentageSplit sometimes fails because it was not well
balanced (>=15%). But as per logs, it just exceeded 1 or 2 requests
like:

https://prow.knative.dev/view/gcs/knative-prow/logs/ci-knative-serving-istio-1.4-no-mesh/1255352174985089024
https://prow.knative.dev/view/gcs/knative-prow/logs/ci-knative-serving-kourier-stable/1255354213559439360
https://prow.knative.dev/view/gcs/knative-prow/logs/ci-knative-serving-kourier-stable/1255022088574472192

e.g.
```
path_test.go:233: Header path-and-percentage-split-sxzgzgfz got: 34 times, want in [35, 65] range
path_test.go:233: Header path-and-percentage-split-oyrvlgmf got: 66 times, want in [35, 65] range
```

So this patch relaxes the tolerance to 20%.