Fix flake in TestPercentage
This patch is same fix with https://github.com/knative/serving/pull/7900.

It increases the margin to 10% from 5% in TestPercentage.
The current 5% margin is not enough to pass the test.
e.g.

https://prow.knative.dev/view/gcs/knative-prow/logs/ci-knative-serving-kourier-stable/1259612209693069312
```
    TestPercentage: percentage_test.go:116: Target "percentage-guxfscre" received 42.400000%, wanted 37.000000 +/- 5.000000
```

And this patch changes to test with concurent request, which is also
same with https://github.com/knative/serving/pull/7900.