Rename operation name for activator's proxy span and queue-proxy's span
Currently both activator's proxy span and queue-proxy's proxy span
have same name `proxy`. It is alright to see one trace data.
But it is difficult to find/filter it by operation name.

Hence, for example, if we want to find the trace which _activator is in the
path_, we cannot filter it.