As discussed in TOC meeting rationalize the owners (#8029)
* As discussed in TOC meeting rationalize the owners

- create TOC and WG-LEAD lists and populate with up to date members
- use those aliases to populate serving-wg-leads alias
- use that alias as the top level approver

With expanded powers come expanded responsibilities.

The gentlemen's agreement is that you don't approve sustantial
changes to the part of the codebase that you're not in the OWNERS
file yet.

E.g. Victor would not approve Service API surface change and Dave
won't approve the change to the logic of autoscaling algorithm.

* use only names in aliases