Optimize happy path in breaker using atomics.
This changes the implementation of the breaker's semaphore to be based on just one atomic operation if we **know** there's capacity in the semaphore. It only relies on channel operations if there is no capacity yet to be able to park and wake up workers that want to wait for capacity in the semaphore.

It does change the semantics around error cases quite a bit in that it is not as "safe" as the old implementation is when it comes to misuse. I would say though that that doesn't really matter in our case because the error case is catastrophic anyway and we have fairly high confidence that we always pair an acquire with a release everywhere we use it.