Use stable pod count if it's larger in panic. (#8106)
Currently in panic mode we only use panic deisred pods to keep the scale. But it's not really correct alway.

Presume the following case: stable window 6s, panic window 5s (bad? sure, possible — just as well).
Data: [1000, 0, 0, 0, 0, 1000] and target 10.
PanicConcurr = (0 + 0 + 0 + 0 + 1000)/5 = 200 => 20 pods
Stable = (1000 + 0 + 0 + 0 + 0 + 1000) / 6 = 333.33 -> 34 pods

So in fact if we keep scaling to panic metric we'll underprovision for stable window.

Thus, this PR picks larger of the two for scaling decisions.
/hold