chore: Hardened cleanup rm expression in update-deps.sh (#8084)
A `rm -rf $(find . ...)` is quite disastrous when you have
directories that contain spaces
(not uncommon on Windows eg, or someone prepared a "w00t /" file for you in your external dependencies).
Also `-r` is not really needed here as no directories are to be removed.

Instead use the idiom `find -print0 | xargs -0` to probably deal with
filenames that have spaces.