Optimise initial startup readiness probing
The probe response body is completely ignored in non-success cases.  Since we
aggressively poll the server up to 40 times a second on startup, it's nice to
shave off a little bit of network stack load if we can. Additionally, the
Golang default transport disables connection re-use via keep-alives if the
response body is not closed _and_ read to EOF, so add an io.Copy to
ioutil.Discard to avoid a lot of unnecessary connection creation overhead at
startup-time.

Also, use queue.Name constant rather than aliveBody since this is the constant
actually used in the corresponding check in revision_backends.go.