Replace hooks with active wais
We have seen in many tests, that hooks are not very reliable (including in this test where we had 2 hooks objects
and now we are down to to one. In many places we actively wait for the actual situation to happen, and then wait for the
hook to succeed.
But it doesn't make sense.
Hooks are also prone to registration ordering, etc.
So get rid of he hooks in the sks tests.