Fix inconsistency in TestContainerErrorMsg.
By only checking that the service is not ready, we don't know if it's the configuration or the route that caused it to be failed. Therefore, we cannot rely on the route being failed statically, without a loop.

This checks in the initial check if the service is failed and makes sure both configuration and route are failed as well. After that, all checks can be transformed to non-retrying checks.