Fix race condition between Revision and PodAutoscaler readiness.
This fixes a race condition where the revision reconciler might never be able to observe a PodAutoscaler while it's ready and only see it non-ready if the PA scales down to 0 before the revision reconciler saw PA's readiness.

While the desire is to make PA's readiness more fitting with K8s APIs in general (i.e. make it be ready even if scaled to 0) this at least fixes the major painpoint we're having with the current state.