Make sorting consistent in GC test.
One of the tests ("keep recent lastPinned") failed fairly often because it used two revisions with the same LastPinned timestamp. Sorting a slice is unstable so in a lot of occasions, the given order was flipped and revision 5555 was tried to be deleted, even though it shouldn't even have been considered to be deleted.

Note: sort.SliceStable doesn't help because the incoming list's order is already non-consistent.