Bump activator terminationGracePeriod to match MaxRevisionTimeout (#8623)
The previous default could lead to active in-flight requests get killed
half-way if the user increased RevisionTimeoutSeconds (since
DefaultMaxRevisionTimeoutSeconds is 10 minutes).

Note that an operator could still increase max-revision-timeout-seconds
in the defaults config map, in which case this value would be too low
again, but at least this makes the out-of-the-box behaviour reasonable.

Added a comment to defaults configmap to ensure an operator knows to
bump this timeout when increasing max-revision-timeout-seconds.