Accept protobuf encoded messages in the statserver. (#8626)
* Accept protobuf encoded messages in the statserver.

This further improves the efficiency of the statserver by allowing it to receive a slice of StatMessages, binary-encoded via protobufs. The protobufs are already quite an improvement over using JSON, both in terms of density and serdes performance. Collapsing all stats into a single message further reduces the overhead of the surrounding protocols, making for a much more efficient protocol overall.

* Dropped unnecessary nil check.

* Change log message and add t.Helper() calls.

* Fix comment.