Introduce a generic getter that gives all the pods broken down by state  (#8760)
* Introduce a generic getter that gives all the pods broken down by state

Currently in KPA we make 3 calls iterating all the pods 3 times (and 3 list calls, which are mostly in memory copying and slice allocations).
With this we can reduce this to 1 (which would come in the next PR).
I didn't add specific tests, since I implemented the other functions via this one and have virtually the same code coverage

* rm