Optimize pod counting in KPA (#8762)
* Introduce a generic getter that gives all the pods broken down by state

Currently in KPA we make 3 calls iterating all the pods 3 times (and 3 list calls, which are mostly in memory copying and slice allocations).
With this we can reduce this to 1 (which would come in the next PR).
I didn't add specific tests, since I implemented the other functions via this one and have virtually the same code coverage

* rm

* Change KPA to use the new helper method.

- simplify the code to get rid of 3 iterations and replace it with 0, when counting pods
- make just one List call as well.
- this highlighted a problem, which we might want to solve separately: when sks is not ready we kept number of pods to be 0, even if it's factually > 0. This is necessary, I think to keep status from becoming active/ready. I've added a TODO to deal with it, since I am unsure about the best way to attack it right now.