Move stat reporting out of the locked section. (#8796)
Even though it currently shouldn't be an issue as the stat reporter in the activator reads as quickly as possible, it could very well become an issue with only small changes. This makes sure that the entire reporter will not be blocked by the channel not being read from quick enough. More specifically: The stat generating parts (i.e. the requests) won't block.