A few more nits (#8793)
* param types

* duplicate

validation set

int32

tweak array copy

tweak array copy

Fix network probing for kingress conformance. (#8764)

Much of kingress conformance uses the "runtime" image to test things because it
echos back headers and other fun stuff.  However, to make this work without the
queue-proxy it has to properly respond to network probes itself as well.  There
was already a probe handler registered at `/healthz` that looked for the kubelet
probe header and replied with a 400 when that wasn't present.  So when we added
`network.ProbePath` to the network probing we actually make things follow a
different request path through the runtime image, and started serving 400s.

You might be wondering why probing didn't fail.

Well, a while back we added logic to the prober to abandon requests that don't
fail with very specific status codes.  If it is a 404 or a 503 we keep trying.
When we get a 400; however, we abandon the probing and return success.

I've been debugging things with an instrumented copy of the probing logic, and I
will try to upstream that in a subsequent PR, but I wanted to land the fix first.

Fix a few nits (#8768)

* param types

* duplicate

* validation set

* int32

* tweak array copy

* tweak array copy

Fix script to get leader in wait_for_leader_controller (#8773)

This patch adds break line to "grep" correctly controller's leader name.

Fail test immediately when important setup failed (#8770)

This patch replaces failed=1 with fail_test in the setup phase in tests.

validation, inline

* add a comment

* keep known labels at the top

* inline objectmeta

* update WithLabel test func

* revert service case

* fix name