Pod random shuffle for scraping, pt I
In this first part of pod random shuffling, we are introducing pod filters and transformers
that permit apply various filtering and final transformation to the pod list.
The current get pod IPs by age is simply re-implemented using this new construct.

In part II, we will be introducing a possibility to cutoff young pods, if possible, and then random shuffling
the rest, rather than sorting them to increase diversity of pods scraped in the metric collector.