Make sure we only return ready pods
When returning pods to scrape for metrics we were returning only running pods.
Which is a necessary, but not satisfactory condition, since a pod might still fail readiness probes.
So add that condition to the list of conditions we check before retruning pods.
It worked mostly so far because our test images are fast to start, so there's not much
time between liveness and readiness spent getting pod ready.
In real life that might not be that rosy.