Make SKS global resync test more consistent.
This test used to rely on a number of endpoints being created by a first reconcilation but didn't actually specify how many endpoints it actually was waiting for. This resulted in intermittent failures being possible.

This makes the test more consistent in that it waits for the reconcilation of all resources to be done statically by awaiting for their observed generation to be bumped.