Ensure when we fail to progress we actually fail revision.
See #8922 for setup.
Without this change the Revision is stuck forever in unknown even after we scaled to zero, while in reality
we failed to progress during the progress deadline (and additional 10s).
Unfortunaly, when we scale to 0, the deployment status switches to True, since scale to zero is a successful
progression of events in the deployment state machine.
But PA has the initial scale attained status, which we can neatly utilize:

 1. if PA is ready=false
 1. if PA never saw initial scale

we can infer the deployment failed to progress to initial state either and thus
we can mark the deployment as failed with ProgressDeadlineExceeded
status.