Make sure we don't override existing message on failed resources. (#8952)
It is easy to envision that PA would fail with this message in addition to deployment failing (e.g. with container pull error message).
Usually deployment failures happen much earlier in time, but overtime PA would fail as well and we'd override the message,
which is not strictly right (not totally wrong either, but still).