Adjust the ContainerExiting test. (#8972)
There is a race to observe the ContainerHealthy condition as the sole failing subcondition of the Revision before the initial scale condition also starts to report problems and (because it is later) overwrites the reason and message on Ready itself.

This relaxes the test to check that the Revision is unhealthy and that at least one of its subconditions reports the desired message.