Allow configuring activator max idle connections (#9027)
* Allow configuring activator max idle connections

Inject the transport into activatorHandler and allow configuring its
default maxIdleConns using a flag. As a bonus, this means the tests for
activatorHandler no longer need to do casts and set the private
transport field.

Using a flag rather than a configmap here because maxIdleConns can't be
easily changed dynamically so we really need the value on startup.

* Add maxIdleConnsPerHost flag too

* Use env vars instead of flag