Cleanup secret accessor tests.
I took a peek at this due to a flake in TestReconcileSecretCreate (link below),
where it timed out waiting on the informer to see a creation.  I cleaned up a
variety of issues while I was looking at this, and to avoid future flakes like
this I changed the test to use the client instead of the lister cache for the
ultimate check, which need not wait as it should be immediately present (and
captures the thrust of the test).

Flake: https://prow.knative.dev/view/gcs/knative-prow/logs/ci-knative-serving-continuous/1295299577284399104